// ui/src/generated/VersionInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { VersionKind } from "./VersionKind";

export type VersionInfo = { id: string, kind: VersionKind, name: string | null, created_ms: number, 
/**
 * Size of the stored document in bytes.
 */
size: number, };